compat-checker: clearer logic and refactoring repeats
