+++ ui/src/channels/ChannelTitleButton.tsx
@@ -1,13 +1,11 @@
 import cn from 'classnames';
 import CaretLeft16Icon from '@/components/icons/CaretLeft16Icon';
 import { useIsMobile } from '@/logic/useMedia';
-import { getFlagParts, isTalk, nestToFlag } from '@/logic/utils';
+import { isTalk } from '@/logic/utils';
 import { useGroupChannel } from '@/state/groups';
 import { Link } from 'react-router-dom';
-import { useConnectivityCheck } from '@/state/vitals';
-import { useChannel } from '@/state/channel/channel';
 import ChannelIcon from './ChannelIcon';
-import HostConnection from './HostConnection';
+import ChannelHostConnection from './ChannelHostConnection';
 
 interface ChannelTitleButtonProps {
   flag: string;
@@ -20,12 +18,7 @@ export default function ChannelTitleButton({
 }: ChannelTitleButtonProps) {
   const isMobile = useIsMobile();
   const channel = useGroupChannel(flag, nest);
-  const [, chFlag] = nestToFlag(nest);
-  const { ship } = getFlagParts(chFlag);
   const BackButton = isMobile ? Link : 'div';
-  const { data } = useConnectivityCheck(ship || '');
-  const chan = useChannel(nest);
-  const saga = chan?.saga || null;
 
   function backTo() {
     if (isMobile && isTalk) {
@@ -56,7 +49,7 @@ export default function ChannelTitleButton({
           )}
         >
           <span className="line-clamp-1">{channel?.meta.title}</span>
-          <HostConnection ship={ship} saga={saga} status={data?.status} />
+          <ChannelHostConnection nest={nest} />
         </div>
         <span className="line-clamp-1 w-full break-all text-sm text-gray-400">
           {channel?.meta.description}
