+++ ui/src/groups/ChannelsList/ChannelManagerHeader.tsx
@@ -1,18 +1,14 @@
 import { useIsMobile } from '@/logic/useMedia';
 import {
   useAmAdmin,
-  useGroup,
   useGroupCompatibility,
   useRouteGroup,
 } from '@/state/groups';
 import cn from 'classnames';
-import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
-import HostConnection from '@/channels/HostConnection';
-import { getFlagParts } from '@/logic/utils';
-import { useConnectivityCheck } from '@/state/vitals';
 import Tooltip from '@/components/Tooltip';
 import ChannelsListSearch from './ChannelsListSearch';
+import GroupHostConnection from '../GroupHostConnection';
 
 interface ChannelManagerHeaderProps {
   addSection: () => void;
@@ -25,14 +21,7 @@ export default function ChannelManagerHeader({
   const flag = useRouteGroup();
   const isAdmin = useAmAdmin(flag);
   const isMobile = useIsMobile();
-  const host = getFlagParts(flag).ship;
-  const { data } = useConnectivityCheck(host);
-  const { saga, text, compatible } = useGroupCompatibility(flag);
-  const hasIssue =
-    !compatible ||
-    (data?.status &&
-      'complete' in data.status &&
-      data.status.complete !== 'yes');
+  const { text, compatible } = useGroupCompatibility(flag);
 
   return (
     <div className="my-4 flex w-full flex-col justify-between space-y-2 sm:flex-row sm:items-center sm:space-x-2">
@@ -64,14 +53,11 @@ export default function ChannelManagerHeader({
       ) : null}
       <div className="flex flex-wrap items-center gap-2">
         <ChannelsListSearch className="w-full flex-1 md:w-[300px]" />
-        {hasIssue && (
-          <HostConnection
-            className={cn(isAdmin && 'order-first')}
-            ship={host}
-            status={data?.status}
-            saga={saga}
-          />
-        )}
+        <GroupHostConnection
+          className={cn(isAdmin && 'order-first')}
+          flag={flag}
+          hideIfConnected
+        />
       </div>
     </div>
   );
