+++ ui/src/groups/GroupActions.tsx
@@ -9,12 +9,7 @@ import { Link, useLocation, useNavigate } from 'react-router-dom';
 import EllipsisIcon from '@/components/icons/EllipsisIcon';
 import useIsGroupUnread from '@/logic/useIsGroupUnread';
 import UnreadIndicator from '@/components/Sidebar/UnreadIndicator';
-import {
-  citeToPath,
-  getFlagParts,
-  getPrivacyFromGroup,
-  useCopy,
-} from '@/logic/utils';
+import { citeToPath, getPrivacyFromGroup, useCopy } from '@/logic/utils';
 import {
   useAmAdmin,
   useGang,
@@ -22,9 +17,6 @@ import {
   useGroupCancelMutation,
 } from '@/state/groups';
 import ActionMenu, { Action } from '@/components/ActionMenu';
-import { Saga } from '@/types/groups';
-import { ConnectionStatus } from '@/state/vitals';
-import HostConnection from '@/channels/HostConnection';
 import { useIsMobile } from '@/logic/useMedia';
 import VolumeSetting from '@/components/VolumeSetting';
 import {
@@ -32,6 +24,7 @@ import {
   useDeletePinMutation,
   usePinnedGroups,
 } from '@/state/pins';
+import GroupHostConnection from './GroupHostConnection';
 
 const { ship } = window;
 
@@ -102,8 +95,6 @@ type GroupActionsProps = PropsWithChildren<{
   open?: boolean;
   onOpenChange?: (open: boolean) => void;
   flag: string;
-  saga?: Saga | null;
-  status?: ConnectionStatus;
   triggerDisabled?: boolean;
   className?: string;
 }>;
@@ -113,8 +104,6 @@ const GroupActions = React.memo(
     open,
     onOpenChange,
     flag,
-    saga,
-    status,
     triggerDisabled,
     className,
     children,
@@ -160,19 +149,12 @@ const GroupActions = React.memo(
       });
     }
 
-    if (saga && isMobile) {
+    if (isMobile) {
       actions.push({
         key: 'connection',
         keepOpenOnClick: true,
         containerClassName: '!p-0 mb-4',
-        content: (
-          <HostConnection
-            ship={getFlagParts(flag).ship}
-            status={status}
-            saga={saga}
-            type="row"
-          />
-        ),
+        content: <GroupHostConnection flag={flag} type="row" />,
       });
     }
